[FIX] base: properly renegerate assets in debug=assets
Recently, the debug=assets mode was changed to use a bundle with
sourcemap instead of serving static files.  Therefore, when someone
modifies a javascript file, refreshing the browser does not load the new
code anymore.

We did not notice that because this is only an issue when the server is
not started in dev mode.  In dev=all (or xml), the assets code does not
use a cache and checks everytime the file checksum number to determine
if a new bundle needs to be done.

It is probably not common for developers to modify source code
without restarting the server and/or using dev=all, but it is a case
that is easy to handle anyway: we just need to make sure that in
debug=assets, the assets code does not use the cache and checks directly
the files checksum.

closes odoo/odoo#66460

Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>