[FIX] web: prevent no content helper on delete/archive list records
before this commit: when deleting or archiving or unarchive all records of last
page in listview user is displayed with NoContentHelper while we still have
other records in previous page, user should be moved to previous page if all
records of last page are deleted.

after this commit: if user deletes or archives or unarchives all records of
last page in list view then user is moved to previous page.

task-2446911

closes odoo/odoo#74091

X-original-commit: a704c7481559c31a66e7b57860b16173072a0300
Signed-off-by: Simon Genin (ges@odoo) <ges@odoo.com>