[FIX] web: race condition in crash manager test
Since commit ee6248cb4e88a6921a58ea32f01d9815df9f513c, the error dialog
has an additional JS lib dependency, which means that it may needs to
perform a request before it opens up. However, a test in the
crashmanager was only waiting for a next tick, which is possibly too
short for a network request.

So, depending on the network speed (and on the test order), this test
could fail. To fix it, we simply make sure that the test also wait for
the library to be loaded

closes odoo/odoo#66546

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>