[FIX] base: fix error message when creating new currency rate
When creating a new currency from the list view, the currency was given 1.0
as rate since this was the field's default value.
The "onchange" on rate would then trigger. If 1.0 is more than 20% different
than the latest rate, a warning was popped. This warning prevented the creation of the new rate.
Meaning one could not create a new currency rate if the latest rate > 1.25 or < 0.833.

This is fixed by setting the latest rate as default value.

closes odoo/odoo#66798

Task: 2450747
X-original-commit: 46a49c6a629b2b60569449b8761b0380ab9a3dc4
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: jbw-odoo <jbw-odoo@users.noreply.github.com>