[FIX] hr_contract: Include archived employees on _get_all_contracts
Purpose
=======

When we want to retrieve all the contract (running for instance), we
don't care about the fact that the employee is archived or not.

For example when we generate the work entries to generate the payslips,
we actually pay the employee, even if he's archived (which is the
normal flow).

closes odoo/odoo#66836

X-original-commit: 5f7a01a8c72c7e8ae4698416d91cd80ba526a598
Related: odoo/enterprise#16677
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>