[IMP] account_edi: added hook when retrying EDI & reword some text
In some cases, non-generic operation needs to be done when retrying to process and edi.document.

For example : for web-services with two steps, if there is an error we need to restart the two steps (and therefore reset the internal state of the edi to the first step, typically by deleting the transaction_id). In case of a warning, we want to retry the second step. This can only be done in the non-generic part of the code, so a hook is added in `account.move`.