[IMP] sale_timesheet: remove SO if the partner is not the same
Before this commit, when the user creates a task in a project in which
the partner is defined, the partner in the task is the one in the
project. If the user changes the partner in the task, then the SO in
this task should not be the one in the project since the partner is
different.

This commit removes the SO if the partner in the task is different than
the one in the SO.

task-2442683

closes #66181