[IMP] account_edi: Replace high velocity CRON by CRON triggers
So, instead of waiting for the cron every hour, we create
a cron trigger, which is created/activated every time
we post/cancel an invoice.  When the CRON worker is available,
it can take the job immediately.

To be sure that we don't miss any crons or that events
might be needed that are not triggered, we still launch the
cron every day.

A test was added to check that the crons are correctly
triggered.

closes odoo/odoo#67088

Signed-off-by: Josse Colpaert <jco@openerp.com>